Reset external MIDI device keys and pedals on pause
and don't send any further MIDI messages to them until in active playback. This removes unnecessary wear and tear on Disklavier mechanisms, which otherwise can have pedal solenoids stuck on for long periods of time when paused, or rapidly fluttering while seeking in non-playback mode, which which wears out the equipment faster and necessitates more maintenance. A side effect of this change is that the status of the keys and pedals of the external MIDI devices can become out of sync with the visualized pedal and keyboard status in the app while paused, but the two will "sync up" as soon as playback resumes and the first new pedal or keypress event occurs, and in any case it's a tradeoff worth making to avoid damaging expensive physical equipment.